core/telemetry: move requestid to pkg directory
