+++ internal/httputil/client_test.go
@@ -6,7 +6,7 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/pomerium/pomerium/internal/telemetry/requestid"
+	"github.com/pomerium/pomerium/pkg/telemetry/requestid"
 )
 
 func TestDefaultClient(t *testing.T) {
